move `IndexerFileWatcher` to a separate file
